requirements.txt
nexmo==1.4.0
mysql_connector_repackaged==0.3.1
pyaudio==0.2.11
setuptools_svn==1.2
simplejson==3.10.0
Flask==0.11.1
